+++ src/database/tests/Stubs/ModelObserverStub.php
@@ -1,0 +1,23 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
+
+namespace HyperfTest\Database\Stubs;
+
+use Hyperf\Database\Model\Events\Updating;
+
+class ModelObserverStub
+{
+    public function updating(Updating $event)
+    {
+        $event->getModel()->foo = 'bar';
+    }
+}