+++ src/grpc-server/src/Server.php
@@ -13,19 +13,11 @@ declare(strict_types=1);
 namespace Hyperf\GrpcServer;
 
 use Hyperf\Contract\ConfigInterface;
-use Hyperf\Dispatcher\HttpDispatcher;
 use Hyperf\GrpcServer\Exception\Handler\GrpcExceptionHandler;
 use Hyperf\HttpServer\Server as HttpServer;
-use Psr\Container\ContainerInterface;
 
 class Server extends HttpServer
 {
-    public function __construct(ContainerInterface $container)
-    {
-        $this->container = $container;
-        $this->dispatcher = $container->get(HttpDispatcher::class);
-    }
-
     public function initCoreMiddleware(string $serverName): void
     {
         $this->serverName = $serverName;