+++ src/json-rpc/src/ConfigProvider.php
@@ -22,8 +22,6 @@ class ConfigProvider
     {
         return [
             'dependencies' => [
-                TcpServer::class => TcpServerFactory::class,
-                HttpServer::class => HttpServerFactory::class,
                 DataFormatter::class => DataFormatterFactory::class,
             ],
             'commands' => [