+++ src/testing/tests/Stub/Exception/Handler/FooExceptionHandler.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+/**
+ * This file is part of Hyperf.
+ *
+ * @link     https://www.hyperf.io
+ * @document https://doc.hyperf.io
+ * @contact  group@hyperf.io
+ * @license  https://github.com/hyperf-cloud/hyperf/blob/master/LICENSE
+ */
+
+namespace HyperfTest\Testing\Stub\Exception\Handler;
+
+use Hyperf\ExceptionHandler\ExceptionHandler;
+use Hyperf\HttpMessage\Stream\SwooleStream;
+use Psr\Http\Message\ResponseInterface;
+use Throwable;
+
+class FooExceptionHandler extends ExceptionHandler
+{
+    public function handle(Throwable $throwable, ResponseInterface $response)
+    {
+        $data = ['code' => $throwable->getCode(), 'message' => $throwable->getMessage()];
+
+        return $response->withBody(new SwooleStream(json_encode($data)));
+    }
+
+    public function isValid(Throwable $throwable): bool
+    {
+        return true;
+    }
+}